Generate NewValueIterator
